Fixed Multifreqdetector and test.
git-svn-id: http://mobicents.googlecode.com/svn/trunk@7408 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ servers/media/core/server-impl/src/main/java/org/mobicents/media/server/impl/resource/tone/MultiFreqToneGeneratorImpl.java
@@ -23,7 +23,7 @@ public class MultiFreqToneGeneratorImpl extends AbstractSource implements MultiF
 	private double dt;
 	private int pSize;
 
-	private short A = Short.MAX_VALUE / 2;
+	private short A = (short)(Short.MAX_VALUE /2);
 
 	private int volume = 0;
